+++ hardware/scripts/gen_trace.py
@@ -718,7 +718,25 @@ def perf_metrics_to_csv(perf_metrics: list, filename: str):
         'stall_raw_lsu',
         'stall_raw_acc',
         'stall_lsu',
-        'stall_acc']
+        'stall_acc',
+        'seq_loads_local',
+        'seq_loads_global',
+        'itl_loads_local',
+        'itl_loads_global',
+        'seq_latency_local',
+        'seq_latency_global',
+        'itl_latency_local',
+        'itl_latency_global',
+        'snitch_load_latency',
+        'snitch_load_region',
+        'snitch_load_tile',
+        'snitch_store_region',
+        'snitch_store_region',
+        'snitch_store_tile',
+        'seq_stores_local',
+        'seq_stores_global',
+        'itl_stores_local',
+        'itl_stores_global']
     for key in keys:
         if key not in known_keys:
             known_keys.append(key)
@@ -728,7 +746,7 @@ def perf_metrics_to_csv(perf_metrics: list, filename: str):
         if write_header:
             dict_writer.writeheader()
         dict_writer.writerows(perf_metrics)
-    print('Wrote performance metrics to %s\n' % filename)
+    print('\nWrote performance metrics to %s\n' % filename)
 
 # -------------------- Main --------------------
 
@@ -776,9 +794,12 @@ def main():
     if args.csv is not None:
         csv_file = args.csv[0]
     path, filename = os.path.split(args.infile.name)
-    core_id = re.search(r'(\d+)', filename)
-    if core_id:
-        core_id = int(core_id.group(1))
+    core_id_hex = re.search(r'(0x[0-9a-fA-F]+)', filename)
+    core_id_dec = re.search(r'([\d]+)', filename)
+    if core_id_hex:
+        core_id = int(core_id_hex.group(1), 16)
+    elif core_id_dec:
+        core_id = int(core_id_dec.group(1))
     else:
         core_id = -1
     # Prepare stateful data structures
@@ -831,15 +852,16 @@ def main():
     # Add metadata
     for sec in perf_metrics:
         sec['core'] = core_id
+    # Emit metrics
+    print('\n## Performance metrics')
+    for idx in range(len(perf_metrics)):
+        print('\n' + fmt_perf_metrics(perf_metrics, idx, not args.allkeys))
+        perf_metrics[idx]['section'] = idx
     # Write metrics to CSV
     if csv_file is not None:
         if os.path.split(csv_file)[0] == '':
             csv_file = os.path.join(path, csv_file)
         perf_metrics_to_csv(perf_metrics, csv_file)
-    # Emit metrics
-    print('\n## Performance metrics')
-    for idx in range(len(perf_metrics)):
-        print('\n' + fmt_perf_metrics(perf_metrics, idx, not args.allkeys))
     # Check for any loose ends and warn before exiting
     seq_isns = len(fseq_info['fseq_pcs']) + len(fseq_info['cfg_buf'])
     unseq_left = len(fseq_info['fpss_pcs']) - len(fseq_info['fseq_pcs'])