+++ scripts/tracevis.py
@@ -16,6 +16,7 @@
 import re
 import os
 import sys
+from functools import lru_cache
 import argparse
 
 has_progressbar = True
@@ -28,40 +29,59 @@ except ImportError as e:
 
 
 # line format:
+# Snitch RTL simulation:
 # 101000 82      M         0x00001000 csrr    a0, mhartid     #; comment
 # time   cycle   priv_lvl  pc         insn
+# MemPool RTL simulation:
+# 101000 82      0x00001000 csrr    a0, mhartid     #; comment
+# time   cycle   pc         insn
+# Banshee traces:
+# 00000432 00000206 0005     800101e0  x15:00000064 x15=00000065 # addi ...
+# cycle    instret  hard_id  pc        register                    insn
 
 # regex matches to groups
 # 0 -> time
 # 1 -> cycle
-# 2 -> privilege level
+# 2 -> privilege level (RTL) / hartid (banshee)
 # 3 -> pc (hex with 0x prefix)
 # 4 -> instruction
-# 5 -> args
-LINE_REGEX = r' *(\d+) +(\d+) +([3M1S0U]?) *(0x[0-9a-f]+) ([.\w]+) +(.+)#'
+# 5 -> args (RTL) / empty (banshee)
+# 6 -> comment (RTL) / instruction arguments (banshee)
+RTL_REGEX = r' *(\d+) +(\d+) +([3M1S0U]?) *(0x[0-9a-f]+) ([.\w]+) +(.+)#; (.*)'
+BANSHEE_REGEX = r' *(\d+) (\d+) (\d+) ([0-9a-f]+) *.+ +.+# ([\w\.]*)( +)(.*)'
 
 # regex matches a line of instruction retired by the accelerator
+# 0 -> time
+# 1 -> cycle
 # 2 -> privilege level
-# 3 -> pc (hex with 0x prefix)
-# 4 -> instruction
-# 5 -> args
-ACC_LINE_REGEX = r' +([3M1S0U]?) *(0x[0-9a-f]+) ([.\w]+) +(.+)#'
-
-re_line = re.compile(LINE_REGEX)
-re_acc_line = re.compile(ACC_LINE_REGEX)
+# 3 -> comment
+ACC_LINE_REGEX = r' *(\d+) +(\d+) +([3M1S0U]?) *#; (.*)'
 
 buf = []
 
 
+@lru_cache(maxsize=1024)
+def addr2line_cache(addr):
+    cmd = f'{addr2line} -e {elf} -f -a -i {addr:x}'
+    return os.popen(cmd).read().split('\n')
+
+
 def flush(buf, hartid):
-    global output_file, use_time
+    global output_file
     # get function names
     pcs = [x[3] for x in buf]
-    a2ls = os.popen(
-        f'addr2line -e {elf} -f -a -i {" ".join(pcs)}').read().split('\n')[:-1]
+    a2ls = []
+
+    if cache:
+        for addr in pcs:
+            a2ls += addr2line_cache(int(addr, base=16))[:-1]
+    else:
+        a2ls = os.popen(
+            f'{addr2line} -e {elf} -f -a -i {" ".join(pcs)}'
+        ).read().split('\n')[:-1]
 
     for i in range(len(buf)-1):
-        (time, cyc, priv, pc, instr, args) = buf.pop(0)
+        (time, cyc, priv, pc, instr, args, cmt) = buf.pop(0)
 
         if use_time:
             next_time = int(buf[0][0])
@@ -69,6 +89,9 @@ def flush(buf, hartid):
         else:
             next_time = int(buf[0][1])
             time = int(cyc)
+
+        # Have lookahead time to this instruction?
+        next_time = lah[time] if time in lah else next_time
 
         # print(f'time "{time}", cyc "{cyc}", priv "{priv}", pc "{pc}"'
         #       f', instr "{instr}", args "{args}"', file=sys.stderr)
@@ -82,12 +105,27 @@ def flush(buf, hartid):
         # print(f'pc "{pc}", func "{func}", file "{file}"')
 
         # assemble values for json
-        label = instr
-        cat = instr
-        start_time = time
+        # Doc: https://docs.google.com/document/d/
+        # 1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview
+        # The name of the event, as displayed in Trace Viewer
+        name = instr
+        # The event categories. This is a comma separated list of categories
+        # for the event. The categories can be used to hide events in the Trace
+        # Viewer UI.
+        cat = 'instr'
+        # The tracing clock timestamp of the event. The timestamps are provided
+        # at microsecond granularity.
+        ts = time
+        # There is an extra parameter dur to specify the tracing clock duration
+        # of complete events in microseconds.
         duration = next_time - time
-        # print(f'"{label}" time {time} next: {next_time}'
-        #       f' duration: {duration}', file=sys.stderr)
+
+        if banshee:
+            # Banshee stores all traces in a single file
+            hartid = priv
+            # In Banshee, each instruction takes one cycle
+            duration = 1
+
         pid = elf+':hartid'+str(hartid)
         funcname = func
 
@@ -100,8 +138,8 @@ def flush(buf, hartid):
         arg_inlined = inlined
 
         output_file.write((
-            f'{{"name": "{label}", "cat": "{cat}", "ph": "X", '
-            f'"ts": {start_time}, "dur": {duration}, "pid": "{pid}", '
+            f'{{"name": "{name}", "cat": "{cat}", "ph": "X", '
+            f'"ts": {ts}, "dur": {duration}, "pid": "{pid}", '
             f'"tid": "{funcname}", "args": {{"pc": "{arg_pc}", '
             f'"instr": "{arg_instr} {arg_args}", "time": "{arg_cycles}", '
             f'"Origin": "{arg_coords}", "inline": "{arg_inlined}"'
@@ -113,35 +151,24 @@ def parse_line(line, hartid):
     # print(line)
     match = re_line.match(line)
     if match:
-        (time, cyc, priv, pc, instr, args) = tuple(
+        (time, cyc, priv, pc, instr, args, cmt) = tuple(
             [match.group(i+1).strip() for i in range(re_line.groups)])
-    # print(match)
-
-    if not match:
-        # match accelerator line with same timestamp as before
-        match = re_acc_line.match(line)
-        if match:
-            (priv, pc, instr, args) = tuple(
-                [match.group(i+1).strip() for i in range(re_acc_line.groups)])
-            # use time,cyc from last line
-            time, cyc = last_time, last_cyc
-        else:
-            return 1
-
-    # print(line)
-    buf.append((time, cyc, priv, pc, instr, args))
-    last_time, last_cyc = time, cyc
+        buf.append((time, cyc, priv, pc, instr, args, cmt))
+        last_time, last_cyc = time, cyc
 
     if len(buf) > 10:
         flush(buf, hartid)
     return 0
 
 
+# Argument parsing
 parser = argparse.ArgumentParser('tracevis', allow_abbrev=True)
 parser.add_argument(
     'elf',
     metavar='<elf>',
     help='The binary executed to generate the traces',
+
+
 )
 parser.add_argument(
     'traces',
@@ -151,19 +178,39 @@ parser.add_argument(
 parser.add_argument(
     '-o',
     '--output',
-    metavar='<trace>',
+    metavar='<json>',
     nargs='?',
     default='chrome.json',
     help='Output JSON file')
+parser.add_argument(
+    '--addr2line',
+    metavar='<path>',
+    nargs='?',
+    default='addr2line',
+    help='`addr2line` binary to use for parsing')
 parser.add_argument(
     '-t',
     '--time',
     action='store_true',
     help='Use the traces time instead of cycles')
 parser.add_argument(
+    '-b',
+    '--banshee',
+    action='store_true',
+    help='Parse Banshee traces')
+parser.add_argument(
+    '--no-cache',
+    action='store_true',
+    help='Disable addr2line caching'
+    ' (slow but might give better traces in some cases)')
+parser.add_argument(
+    '--overlap-instructions',
+    action='store_true',
+    help='Lookahead for instruction duration and report their full duration')
+parser.add_argument(
     '-s',
     '--start',
-    metavar='<trace>',
+    metavar='<line>',
     nargs='?',
     type=int,
     default=0,
@@ -171,7 +218,7 @@ parser.add_argument(
 parser.add_argument(
     '-e',
     '--end',
-    metavar='<trace>',
+    metavar='<line>',
     nargs='?',
     type=int,
     default=-1,
@@ -183,19 +230,88 @@ elf = args.elf
 traces = args.traces
 output = args.output
 use_time = args.time
+banshee = args.banshee
+addr2line = args.addr2line
+cache = not args.no_cache
+lookahead = args.overlap_instructions
 
-print('elf', elf, file=sys.stderr)
-print('traces', traces, file=sys.stderr)
-print('output', output, file=sys.stderr)
+print('elf:', elf, file=sys.stderr)
+print('traces:', traces, file=sys.stderr)
+print('output:', output, file=sys.stderr)
+print('addr2line:', addr2line, file=sys.stderr)
+print('cache:', cache, file=sys.stderr)
+
+# Compile regex
+if banshee:
+    re_line = re.compile(BANSHEE_REGEX)
+else:
+    re_line = re.compile(RTL_REGEX)
+
+re_acc_line = re.compile(ACC_LINE_REGEX)
+
+
+def offload_lookahead(lines):
+    # dict mapping time stamp of retired instruction to time stamp of
+    # accelerator complete
+    lah = {}
+    searches = []
+    re_load = re.compile(r'([a-z]*[0-9]*|zero) *<~~ Word')
+
+    for line in lines:
+        match = re_line.match(line)
+        if match:
+            (time, cyc, priv, pc, instr, args, cmt) = tuple(
+                [match.group(i+1).strip() for i in range(re_line.groups)])
+            time = int(time) if use_time else int(cyc)
+
+            # register searchers
+            if '<~~ Word' in cmt:
+                if re_load.search(cmt):
+                    dst_reg = re_load.search(cmt).group(1)
+                    pat = f'(lsu) {dst_reg}  <--'
+                    searches.append({'pat': pat, 'start': time})
+                else:
+                    print(f'unsupported load lah: {cmt}')
+
+        # If this line is an acc-only line, get the data
+        if not match:
+            match = re_acc_line.match(line)
+            if match:
+                (time, cyc, priv, cmt) = tuple(
+                    [match.group(i+1).strip()
+                     for i in range(re_acc_line.groups)])
+
+        time = int(time) if use_time else int(cyc)
+
+        # Check for any open searches
+        removes = []
+        for s in searches:
+            if s['pat'] in cmt:
+                lah[s['start']] = time
+                removes.append(s)
+        [searches.remove(r) for r in removes]
+
+    # for l in lah:
+    #     print(f'{l} -> {lah[l]}')
+    return lah
+
+
+lah = {}
 
 with open(output, 'w') as output_file:
     # JSON header
     output_file.write('{"traceEvents": [\n')
 
+    hartid = 0
     for filename in traces:
-        hartid = 0
-        parsed_nums = re.findall(r'\d+', filename)
-        hartid = int(parsed_nums[-1]) if len(parsed_nums) else hartid+1
+        hartid_hex = re.search(r'(0x[0-9a-fA-F]+)', filename)
+        hartid_dec = re.search(r'([\d]+)', filename)
+        if hartid_hex:
+            hartid = int(hartid_hex.group(1), 16)
+        elif hartid_dec:
+            hartid = int(hartid_dec.group(1))
+        else:
+            hartid = hartid+1
         fails = lines = 0
         last_time = last_cyc = 0
 
@@ -203,15 +319,19 @@ with open(output, 'w') as output_file:
             f'parsing hartid {hartid} with trace {filename}', file=sys.stderr)
         tot_lines = len(open(filename).readlines())
         with open(filename) as f:
+            all_lines = f.readlines()[args.start:args.end]
+            # offload lookahead
+            if lookahead:
+                lah = offload_lookahead(all_lines)
             if has_progressbar:
                 for lino, line in progressbar.progressbar(
-                        enumerate(f.readlines()[args.start:args.end]),
+                        enumerate(all_lines),
                         max_value=tot_lines):
                     fails += parse_line(line, hartid)
                     lines += 1
             else:
                 for lino, line in enumerate(
-                        f.readlines()[args.start:args.end]):
+                        all_lines):
                     fails += parse_line(line, hartid)
                     lines += 1
             flush(buf, hartid)